requirements.txt
fastapi
uvicorn
asyncpg
python-dotenv
pydantic
httpx
psycopg2-binary
